Fixed setEventData(data) to setEventDat(data.event)
