+++ src/front/js/layout.js
@@ -5,7 +5,7 @@ import { BackendURL } from "./component/backendURL";
 
 import { Home } from "./pages/home";
 import { Events } from "./pages/events";
-import { Single } from "./pages/single";
+import Single from "./pages/single"; // Updated import
 import injectContext from "./store/appContext";
 
 import { Navbar } from "./component/navbar";
@@ -14,11 +14,9 @@ import { LogIn } from "./pages/login";
 import { SignUp } from "./pages/signUp";
 import { CreateEvent } from "./pages/createEvent";
 import { Contact } from "./pages/contact";
+import EventSingle from "./component/EventSingle";
 
-//create your first component
 const Layout = () => {
-    //the basename is used when your project is published in a subdirectory and not in the root of the domain
-    // you can set the basename on the .env file located at the root of this project, E.g: BASENAME=/react-hello-webapp/
     const basename = process.env.BASENAME || "";
 
     if(!process.env.BACKEND_URL || process.env.BACKEND_URL == "") return <BackendURL/ >;
@@ -31,15 +29,13 @@ const Layout = () => {
                     <Routes>
                         <Route element={<Home />} path="/" />
                         <Route element={<Events />} path="/events" />
-                        <Route element={<Single />} path="/single" />
+                        <Route element={<Single />} path="/single/:id" /> {/* Updated route */}
                         <Route element={<Contact />} path="/contact" />
                         <Route element={<LogIn />} path="/login" />
                         <Route element={<SignUp />} path="/sign-up" />
                         <Route element={<CreateEvent />} path="/create-event" />
-                        {/*<Route element={<Private />} path="/private" />*/}
-
-                        {/* <Route element={<Single />} path="/single/:theid" /> */}
                         <Route element={<h1>Not found!</h1>} />
+                        <Route element={<EventSingle />} path="/single-event" />
                     </Routes>
                     <Footer />
                 </ScrollToTop>
